QOL and bugfix and better test graph

+++ tools/download_results.py
@@ -7,7 +7,8 @@ def main(mode: str, from_index: int, to_index: int, connection_str: str):
         [f"{mode}_{i}.log" for i in range(from_index, to_index)]
         
     for file in files:
-        os.remove(f"./{mode}/{file}")
+        if os.path.exists(f"./{mode}/{file}"):
+            os.remove(f"./{mode}/{file}")
         
     subprocess.run(f"scp {connection_str}:~/{{{",".join(files)}}} ./{mode}/")
